Revert "Merge branch 'master' into FollowingPokemon_Gen1-3"
This reverts commit 63b2bd9a44c4cf5faa561160dec788573f819a7b, reversing
changes made to a716f522f3599ecf073c7e302e3c72dbaac053c3.

+++ src/scrcmd.c
@@ -1179,7 +1179,7 @@ bool8 ScrCmd_setobjectmovementtype(struct ScriptContext *ctx)
 
 bool8 ScrCmd_createvobject(struct ScriptContext *ctx)
 {
-    u8 graphicsId = ScriptReadByte(ctx);
+    u16 graphicsId = ScriptReadByte(ctx);
     u8 virtualObjId = ScriptReadByte(ctx);
     u16 x = VarGet(ScriptReadHalfword(ctx));
     u32 y = VarGet(ScriptReadHalfword(ctx));
